% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_generics.R
\name{boot_tidy}
\alias{boot_tidy}
\title{Bootstrapped dataframe with estimates from each sample.}
\usage{
boot_tidy(data, times = 25, strata = NULL, apparent = FALSE, ..f,
  ..., tidy.args = list())
}
\arguments{
\item{data}{Dataframe (or tibble) from which variables are to be taken.}

\item{times}{The number of bootstrap samples.}

\item{strata}{A variable that is used to conduct stratified sampling. When not \code{NULL}, each bootstrap sample is created within the stratification variable.}

\item{apparent}{A logical. Should an extra resample be added where the analysis and holdout subset are the entire data set. This is required for some estimators used by the \code{summary} function that require the apparent error rate.}

\item{..f}{A function, or function name as a string.}

\item{...}{Arguments to function.

These dots support \link{tidy-dots} features.}

\item{tidy.args}{A list of arguments to be used in the relevant \code{S3} method.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with information about model components.
}
\description{
Bootstrapped dataframe with estimates from each sample.
}
\note{
For available methods, see-
\url{https://indrajeetpatil.github.io/broomExtra/articles/available_methods.html}
}
\section{Methods}{


\Sexpr[stage=render,results=rd]{generics:::methods_rd("tidy")}

}

\examples{
set.seed(123)

# example-1: linear model
broomExtra::boot_tidy(
  data = mtcars,
  times = 500,
  ..f = stats::lm,
  formula = mpg ~ wt,
  na.action = na.omit,
  tidy.args = list(conf.int = TRUE, conf.level = 0.50)
)

# example-2: linear mixed-effects model
library(lme4)

broomExtra::boot_tidy(
  data = sleepstudy,
  times = 25,
  ..f = lme4::lmer,
  formula = Reaction ~ Days + (Days | Subject),
  tidy.args = list(effects = "fixed")
)
}
\seealso{
\code{\link{grouped_tidy}}, \code{\link{tidy}}
}
\author{
Indrajeet Patil
}
