% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouped_generics.R
\name{grouped_augment}
\alias{grouped_augment}
\title{Augmented data from grouped analysis of any function that has \code{data}
argument in its function call.}
\usage{
grouped_augment(data, grouping.vars, ..f, ..., augment.args = list())
}
\arguments{
\item{data}{Dataframe (or tibble) from which variables are to be taken.}

\item{grouping.vars}{Grouping variables.}

\item{..f}{A function, or function name as a string.}

\item{...}{Arguments to function.

These dots support \link{tidy-dots} features.}

\item{augment.args}{A list of arguments to be used in the relevant \code{S3} method.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with information about data points.
}
\description{
Augmented data from grouped analysis of any function that has \code{data}
argument in its function call.
}
\note{
For available methods, see-
\url{https://indrajeetpatil.github.io/broomExtra/articles/available_methods.html}
}
\section{Methods}{


\Sexpr[stage=render,results=rd]{generics:::methods_rd("augment")}

}

\examples{
set.seed(123)
# to speed up computation, let's use only 50\% of the data

# linear model
broomExtra::grouped_augment(
  data = dplyr::sample_frac(tbl = ggplot2::diamonds, size = 0.5),
  grouping.vars = c(cut, color),
  formula = price ~ carat - 1,
  ..f = stats::lm,
  na.action = na.omit,
  augment.args = list(se_fit = TRUE)
)

# linear mixed effects model
broomExtra::grouped_augment(
  data = dplyr::sample_frac(tbl = ggplot2::diamonds, size = 0.5),
  grouping.vars = cut,
  ..f = lme4::lmer,
  formula = price ~ carat + (carat | color) - 1,
  control = lme4::lmerControl(optimizer = "bobyqa")
)
}
\seealso{
\code{\link{augment}}, \code{\link{boot_augment}}
}
\author{
\href{https://github.com/IndrajeetPatil}{Indrajeet Patil}
}
