% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouped_tidy.R
\name{grouped_tidy}
\alias{grouped_tidy}
\title{Tidy output from grouped analysis of any function that has \code{data}
argument in its function call.}
\usage{
grouped_tidy(data, grouping.vars, ..f, ..., tidy.args = list())
}
\arguments{
\item{data}{Dataframe (or tibble) from which variables are to be taken.}

\item{grouping.vars}{Grouping variables.}

\item{..f}{A function, or function name as a string.}

\item{...}{Arguments to function.

These dots support \link{tidy-dots} features.}

\item{tidy.args}{A list of arguments to be used in the relevant \code{S3} method.}
}
\description{
Tidy output from grouped analysis of any function that has \code{data}
argument in its function call.
}
\examples{
set.seed(123)

# linear model
grouped_tidy(
  data = mtcars,
  grouping.vars = c(cyl),
  formula = wt ~ am - 1,
  ..f = stats::lm,
  na.action = na.omit,
  tidy.args = list(conf.int = TRUE, conf.level = 0.99)
)
}
\author{
Indrajeet Patil
}
