% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{tidy}
\alias{tidy}
\title{Retrieve tidy dataframe if it exists.}
\usage{
tidy(x, ...)
}
\arguments{
\item{x}{An object to be converted into a tidy \code{\link[tibble:tibble]{tibble::tibble()}}.}

\item{...}{Additional arguments to tidying method.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with information about model components.
}
\description{
Checks if a \code{tidy} method exits for a given object, either in
\code{broom} or in \code{broom.mixed}. If it does, it turn an object into a tidy
tibble, if not, return a \code{NULL}. In case of data frames, a tibble data
frame is returned.
}
\note{
For available methods, see-
\url{https://indrajeetpatil.github.io/broomExtra/articles/available_methods.html}
}
\section{Methods}{


\Sexpr[stage=render,results=rd]{generics:::methods_rd("tidy")}

}

\examples{
set.seed(123)
library(lme4)

# mixed-effects models (`broom.mixed` will be used)
lmm.mod <- lmer(Reaction ~ Days + (Days | Subject), sleepstudy)
broomExtra::tidy(x = lmm.mod, effects = "fixed", exponentiate = TRUE)

# linear model (`broom` will be used)
lm.mod <- lm(Reaction ~ Days, sleepstudy)
broomExtra::tidy(x = lm.mod, conf.int = TRUE)

# in case you enter a dataframe, tidier will return data as a tibble
# note that rownames are preserved but other attributes might be lost
broomExtra::tidy(mtcars)

# unsupported object (the function will return `NULL` in such cases)
broomExtra::tidy(list(1, c("x", "y")))
}
\seealso{
\code{\link{grouped_tidy}}, \code{\link{boot_tidy}}
}
\author{
\href{https://github.com/IndrajeetPatil}{Indrajeet Patil}
}
